pyproject.toml
[tool.poetry]
name = "import-mimes"
version = "0.1.0"
description = ""
authors = ["Travis James <travis@tribehealthsolutions.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.25.2"
python-dotenv = "^1.0.1"
supabase = "^2.4.5"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
